+++ frontend/src/routes/index.tsx
@@ -1,11 +1,12 @@
-import { Navigate, createBrowserRouter, Outlet, useLocation } from 'react-router-dom';
+import { Navigate, createBrowserRouter, Outlet } from 'react-router-dom';
 import Login from '../pages/Login';
 import Dashboard from '../pages/Dashboard';
 import NavBar from '../navigation/NavBar';
 import AuthPage from '../pages/AuthPage';
-import { useAuth } from '../utils/hooks';
+import PublicOnly from './PublicOnly';
+import RequireAuth from './RequireAuth';
 
-export default createBrowserRouter([
+export const router = createBrowserRouter([
   {
     element: (
       <>
@@ -41,27 +42,4 @@ export default createBrowserRouter([
   },
 ]);
 
-function RequireAuth() {
-  const auth = useAuth();
-  const location = useLocation();
-
-  if (!auth.authState) {
-    // Redirect them to the /login page, but save the current location they were
-    // trying to go to when they were redirected. This allows us to send them
-    // along to that page after they login, which is a nicer user experience
-    // than dropping them off on the home page.
-    return <Navigate to="/login" state={{ from: location }} replace />;
-  } else {
-    return <Outlet />;
-  }
-}
-
-function PublicOnly() {
-  const auth = useAuth();
-  const location = useLocation();
-  if (auth.authState) {
-    return <Navigate to="/dashboard" state={{ from: location }} replace />;
-  } else {
-    return <Outlet />;
-  }
-}
+export default router;
